test(trusty-code): prove #2279 verify-gate block-then-retry intercept
The #2279 finish_task verify-gate was proven live/wired but its
block->nudge->retry path had never fired end-to-end (agents always
self-verify before finishing). Add deterministic ScriptedLlm tests that
force a premature finish_task and assert the gate rejects it, injects the
retry nudge (observed in the fed-back tool-role message), and recovers on
retry; plus a negative control sharing the identical armed seed. Extract
the gate tests to a child module to stay under the 1500 SLOC test cap.

Refs #2279
